requirements.txt
streamlit==1.39.0
python-dotenv==1.0.1
openai==1.51.2
langchain==0.3.7
langgraph==0.2.35
amadeus==9.0.0
tenacity==8.5.0
orjson==3.10.7
langchain-huggingface==0.1.0
google-generativeai==0.7.2

